requirements.txt
selenium
selenium-wire
jupyterlab
pandas
numpy
matplotlib
pyyaml
openpyxl
